Move reaction stuff to user/reactions, plotting to deprecated. Consider moving top level gap_rss_iter_fit to its own user subdirectory as well, or maybe another repo?
