+++ deprecated/plotting/maxveit_plottools.py
