+++ deprecated/plotting/normal_modes.py
