+++ deprecated/plotting/plot_2b.py
