+++ deprecated/plotting/plot_ef_correlation.py
