+++ user/reactions/generate/collision.py
