+++ user/reactions/generate/irc.py
