+++ wfl/cli/cli.py
@@ -14,7 +14,6 @@ from pathlib import Path
 import ase.data
 import ase.io
 import click
-import matplotlib.pyplot as plt
 import numpy as np
 import yaml
 
@@ -24,18 +23,11 @@ except ModuleNotFoundError:
     pass
 # noinspection PyProtectedMember
 from ase.io.extxyz import key_val_str_to_dict
-from matplotlib.backends.backend_pdf import PdfPages
 
-from wfl.plotting import reactions_plotting, plot_ef_correlation, plot_2b
-from wfl.plotting import normal_modes
 from wfl.configset import ConfigSet, OutputSpec
-import wfl.generate.collision
 from wfl.generate import vib
-import wfl.generate.radicals
 import wfl.generate.smiles
 import wfl.utils.misc
-from wfl.reactions_processing import trajectory_processing
-from wfl.select import weighted_cur
 import wfl.generate.buildcell
 import wfl.select.by_descriptor
 import wfl.descriptors.quippy
@@ -76,12 +68,6 @@ def subcli_file_operations(ctx):
 @cli.group("processing")
 @click.pass_context
 def subcli_processing(ctx):
-    pass
-
-
-@cli.group("plot")
-@click.pass_context
-def subcli_plot(ctx):
     pass
 
 
@@ -144,99 +130,6 @@ def configs_from_smiles(ctx, smiles, output, info, force):
         print(outputspec)
 
     wfl.generate.smiles.run(outputs=outputspec, smiles=smiles, extra_info=info)
-
-
-@subcli_generate_configs.command('remove-sp3-Hs')
-@click.pass_context
-@click.argument("inputs", nargs=-1)
-@click.option("--outputs", "-o", help="Output filename, see Configset for details", required=True)
-@click.option("--force", "-f", help="force writing", is_flag=True)
-def configs_from_smiles(ctx, inputs, outputs, force):
-    """Removes all sp3 hydrogen atoms for given molecules"""
-
-    verbose = ctx.obj["verbose"]
-
-    inputs = ConfigSet(input_files=inputs)
-    outputs = OutputSpec(output_files=outputs, force=force)
-
-    if verbose:
-        print(inputs)
-        print(outputs)
-
-    wfl.generate.radicals.abstract_sp3_hydrogen_atoms(inputs=inputs, outputs=outputs)
-
-
-@subcli_generate_configs.command("collision")
-@click.option("--fragments", "--fn", default="fragments.xyz",
-              help="Filename of fragments, ase-readable file")
-@click.option("--gap-filename", "--gap", default="GAP.xml", help="Filename of GAP xml file")
-@click.option("--md-dir", "--dir", default="md", help="Directory for calculation")
-@click.option("--velo", default=(0.15, 0.20),
-              help="Velocity parameters: (factor, constant). See Supercollider")
-@click.option("--nsteps", "-n", default=1000, type=click.INT, help="Number of MD steps")
-@click.option("--temp", "-t", default=1000., type=click.FLOAT, help="Temperature in K")
-@click.option("--distance", "-d0", "-d", default=6., type=click.FLOAT,
-              help="Initial distance of molecules")
-@click.option("--collision-kwargs", "--kw", type=click.STRING, required=False,
-              help="Kwargs for collision, overwrites options.")
-@click.pass_context
-def collision(ctx, fragments, gap_filename, md_dir, velo, nsteps, temp, distance, collision_kwargs):
-    verbose = ctx.obj["verbose"]
-    if verbose:
-        print("we are in verbose mode")
-
-    fragment_list = ase.io.read(fragments, ":")
-    md_dir = os.path.abspath(md_dir)
-
-    if not os.path.isdir(md_dir):
-        if verbose:
-            print(f"Creating directory for md: {md_dir}")
-        os.mkdir(md_dir)
-
-    collision_kw = dict(velocity_params=velo, nsteps=nsteps, T=temp, d0=distance)
-    if collision_kwargs is not None:
-        collision_kw.update(key_val_str_to_dict(collision_kwargs))
-
-    wfl.generate.collision.multi_run_all_with_all(
-        fragments=fragment_list, param_filename=os.path.abspath(gap_filename), workdir=md_dir,
-        **collision_kw)
-
-
-@subcli_generate_configs.command("neb-ts-irc")
-@click.pass_context
-@click.argument("seeds", nargs=-1)
-@click.option("--gap-filename", "--gap", default="GAP.xml", help="Filename of GAP xml file")
-@click.option("--do-neb", is_flag=True, help="Calculate NEB between differing relaxed frames")
-@click.option("--do-ts-irc", is_flag=True, help="Calculate TS & IRC on NEBs")
-@click.option("--minim-interval", default=50,
-              help="Interval of trajectory frames for calculation of relaxations from")
-@click.option("--n-pool", "-n", default=None, type=click.INT, help="Number of pooled processes")
-@click.option("--minim-kwargs", "--min-kw", type=click.STRING, required=False,
-              help="Kwargs for NEB")
-@click.option("--neb-kwargs", "--neb-kw", type=click.STRING, required=False, help="Kwargs for NEB")
-@click.option("--ts-kwargs", "--ts-kw", type=click.STRING, required=False, help="Kwargs for TS")
-@click.option("--irc-kwargs", "--irc-kw", type=click.STRING, required=False, help="Kwargs for IRC")
-def trajectory_neb_ts_irc(ctx, seeds, gap_filename, do_neb, do_ts_irc, minim_interval, minim_kwargs,
-                          neb_kwargs,
-                          ts_kwargs, irc_kwargs, n_pool):
-    seeds = [trajectory_processing.xyz_to_seed(fn) for fn in seeds]
-    print(seeds)  # debug
-    calc = (quippy.potential.Potential, "", dict(param_filename=gap_filename))
-
-    if minim_kwargs is not None:
-        minim_kwargs = key_val_str_to_dict(minim_kwargs)
-    if neb_kwargs is not None:
-        neb_kwargs = key_val_str_to_dict(neb_kwargs)
-    if ts_kwargs is not None:
-        ts_kwargs = key_val_str_to_dict(ts_kwargs)
-    if irc_kwargs is not None:
-        irc_kwargs = key_val_str_to_dict(irc_kwargs)
-
-    wfl.generate.collision.post_process_collision(
-        seed=seeds, calc=calc, do_neb=do_neb, do_ts_irc=do_ts_irc, minim_kwargs=minim_kwargs,
-        minim_interval=minim_interval, neb_kwargs=neb_kwargs,
-        ts_kwargs=ts_kwargs, irc_kwargs=irc_kwargs, n_pool=n_pool,
-    )
 
 
 @subcli_file_operations.command("gather")
@@ -366,303 +259,6 @@ def calc_ef_committee(ctx, inputs, prefix, gap_fn, stride, force):
     outputspec.end_write()
 
 
-@subcli_processing.command("max-similarity")
-@click.pass_context
-@click.argument("inputs", nargs=-1)
-@click.option("--train-filename", "--tr", help="Training filename - reference to use",
-              type=click.STRING,
-              required=False)
-@click.option("--cutoff-list", "-c", help="Cutoffs to calculate SOAP with", type=click.STRING,
-              multiple=True,
-              default=("3.0", "6.0"), required=False)
-@click.option("--z-list", "-z",
-              help="Atomic numbers to include in SOAP - incomplete implementation atm.",
-              type=click.INT, multiple=True, default=(1, 6, 8), required=False)
-@click.option("--prefix", "-p", help="Prefix to put on the first", type=click.STRING,
-              default="sim.")
-@click.option("--force", "-f", help="force writing", is_flag=True)
-def calc_max_kernel_similarity(ctx, inputs, force, train_filename, cutoff_list, z_list, prefix):
-    """Calculate maximum kernel similarity with a given training set.
-
-    configset is working in many -> many mode
-    """
-    verbose = ctx.obj["verbose"]
-
-    # apply the prefix to output names
-    outputs = {fn: os.path.join(os.path.dirname(fn), f"{prefix}{os.path.basename(fn)}") for fn in
-               inputs}
-
-    configset = ConfigSet(input_files=inputs)
-    outputspec = OutputSpec(output_files=outputs, force=force)
-    if verbose:
-        print(configset)
-        print(outputspec)
-        sys.stdout.flush()
-
-    # initialisations for calculation
-    frames_train = ase.io.read(train_filename, ":")
-    soap_dict, desc_ref = trajectory_processing.create_desc_ref(frames_train, cutoff_list, z_list)
-    if verbose:
-        print("Calculated SOAP vectors for training set")
-
-    for at in configset:
-        at = trajectory_processing.calc_max_similarity_atoms(at, soap_dict, desc_ref)
-        outputspec.write(at, from_input_file=configset.get_current_input_file())
-
-    outputspec.end_write()
-
-
-@subcli_select_configs.command("weighted-cur")
-@click.pass_context
-@click.argument("inputs", nargs=-1)
-@click.option("--cut-threshold", type=click.FLOAT, default=None,
-              help="Cut each trajectory when combined similarity metric [exp(mean(log(k), "
-                   "over atoms)] falls "
-                   "below threshold.")
-@click.option("--limit", "-l", type=click.FLOAT, required=True,
-              help="Energy limit below which lower weight is given, above it it is linear")
-@click.option("--output", "-o", required=True, help="Output filename, see Configset for details", )
-@click.option("--stride", "-i", type=click.INT, required=False,
-              help="Pass this index to configset globally")
-@click.option("--descriptor", "-d", type=click.STRING, multiple=True,
-              default=(
-              "soap n_max=8 l_max=6 atom_sigma=0.3 n_species=3 species_Z={6 1 8 } cutoff=3.0",),
-              help="SOAP string, species-Z will be added to this!")
-@click.option("--n-select", "-n", type=(click.INT, click.INT), multiple=True,
-              help="(z, num) tuples of how many to select per atomic number, per cutoff instance")
-@click.option("--force", "-f", help="force writing", is_flag=True)
-def select_cur_and_committee(ctx, inputs, output, cut_threshold, limit, descriptor, n_select, force,
-                             stride):
-    """Selection with weighting CUR with energy std of committee of models after cutting with
-    global metric on traj
-
-    Notes
-    -----
-    - This is too specific now, should maybe split the trajectory manipulation and CUR into two
-    and merge tha latter
-    into the general CUR we have here
-    """
-    verbose = ctx.obj["verbose"]
-
-    configset = ConfigSet(input_files=inputs,
-                                default_index=(f"::{stride}" if stride is not None else ":"))
-    if cut_threshold is not None:
-        # cutting by global SOAP metric -- simply recreating the configset with indices calculated
-        new_inputs = []
-
-        for subcfs in configset.group_iter():
-            idx = trajectory_processing.cut_trajectory_with_global_metric(subcfs, cut_threshold)
-            current_fn = configset.get_current_input_file()
-            if verbose:
-                print(f"cutting at index: {idx} on file {current_fn}")
-
-            if stride is None:
-                if idx == -1:
-                    str_index = ":"
-                else:
-                    str_index = f":{idx}"
-            else:
-                if idx == -1:
-                    str_index = f"::{stride}"
-                else:
-                    str_index = f":{idx * stride}:{stride}"
-
-            new_inputs.append((current_fn, str_index))
-
-        # recreate the configset to have the file indices in it
-        configset = ConfigSet(input_files=new_inputs)
-
-    outputspec = OutputSpec(output_files=output, force=force)
-
-    if verbose:
-        print(configset)
-        print(outputspec)
-        sys.stdout.flush()
-
-    z_list = []
-    num_dict = dict()
-    for z, num in n_select:
-        z_list.append(z)
-        num_dict[z] = num
-    if verbose:
-        print("(z, num) to take:", num_dict)
-
-    weighted_cur.selection(configset, outputspec, z_list, descriptor, limit, num_dict)
-    outputspec.end_write()
-
-
-
-@subcli_plot.command("ef-thinpoints")
-@click.pass_context
-@click.argument("inputs", nargs=-1)
-@click.option("--plot-fn", type=click.STRING, default="plot_thinpoints.pdf",
-              help="Filename for plot")
-@click.option("--gap-fn", type=click.STRING, default="GAP.xml", help="Filename of GAP, used for e0")
-@click.option("--ref-energy-key", "-re", type=click.STRING, default="gap_energy",
-              help="Ref energy key")
-@click.option("--gap-energy-key", "-ge", type=click.STRING, default="ref_energy",
-              help="GAP energy key")
-@click.option("--ref-force-key", "-rf", type=click.STRING, default="ref_forces",
-              help="Ref force key")
-@click.option("--gap-force-key", "-gf", type=click.STRING, default="gap_forces",
-              help="GAP force key")
-@click.option("--colors", type=click.STRING, help="Colors dict as str")
-@click.option("--thin-kwargs", type=click.STRING, help="Kwargs for thinpoints")
-def plot_ef_thinpoints(ctx, inputs, plot_fn, gap_fn, colors, thin_kwargs, ref_energy_key,
-                       gap_energy_key,
-                       ref_force_key, gap_force_key):
-    """Plot energy and force correlation with thinpoints
-    """
-    frames = ConfigSet(input_files=inputs)
-    print(frames)
-    e0 = gap_xml_tools.extract_e0(gap_fn)
-    if colors is not None:
-        colors = key_val_str_to_dict(colors)
-    if thin_kwargs is not None:
-        thin_kwargs = key_val_str_to_dict(thin_kwargs)
-
-    plot_ef_correlation.plot(frames=frames, fig_filename=plot_fn, e0=e0,
-                             ref_energy_key=ref_energy_key, ref_force_key=ref_force_key,
-                             gap_energy_key=gap_energy_key, gap_force_key=gap_force_key,
-                             colors=colors, thin_params=thin_kwargs)
-
-
-@subcli_plot.command("2b")
-@click.pass_context
-@click.argument("atomic-numbers", nargs=-1, type=click.INT)
-@click.option("--plot-fn", type=click.STRING, default="plot_2b.pdf", help="Filename for plot")
-@click.option("--gap-fn", type=click.STRING, default="GAP.xml", help="Filename of GAP")
-@click.option("--baseline-fn", type=click.STRING, help="Filename of baseline model")
-@click.option("--baseline-label", type=click.STRING, default="glue",
-              help="label of baseline model for plot")
-@click.option("--baseline-kw", type=click.STRING, help="kwargs for initialising baseline model")
-@click.option("--colors", type=click.STRING, help="Colors dict as str")
-@click.option("--cutoff", type=click.FLOAT, help="Cutoff un to which to calculate binding energies")
-def plot_2b_binding(ctx, atomic_numbers, plot_fn, gap_fn, colors, baseline_fn, baseline_label,
-                    baseline_kw,
-                    cutoff=None):
-    if colors is not None:
-        colors = key_val_str_to_dict(colors)
-
-    gap = (quippy.potential.Potential, "", dict(param_filename=gap_fn))
-    e0 = gap_xml_tools.extract_e0(gap_fn)
-
-    # only create baseline if filename is given
-    if baseline_fn is not None:
-        if baseline_kw is None:
-            baseline_kw = dict()
-            if baseline_label == "glue":
-                baseline_kw["args_str"] = "IP Glue"
-        else:
-            baseline_kw = key_val_str_to_dict(baseline_kw)
-        baseline_calc = quippy.potential.Potential(param_filename=baseline_fn, **baseline_kw)
-    else:
-        baseline_calc = None
-
-    plot_2b.plot_2b_multipage(gap_calculator=gap, baseline_calculator=baseline_calc,
-                              fig_filename=plot_fn, e0=e0,
-                              atomic_numbers=atomic_numbers, cutoff=cutoff,
-                              baseline_label=baseline_label,
-                              colors=colors)
-
-
-@subcli_plot.command("traj-std")
-@click.pass_context
-@click.argument("inputs", nargs=-1)
-@click.option("--plot-fn", type=click.STRING, default="plot.pdf",
-              help="Filename for plot, multi-page")
-@click.option("--data-fn", type=click.STRING, default="data.npy",
-              help="Filename for compressed data storage")
-@click.option("--enforce-rebuild", "-f", is_flag=True,
-              help="Enforce rebuilding of the compressed data")
-def plot_trajectory_std(ctx, inputs, data_fn, enforce_rebuild, plot_fn):
-    """Plots the estimated STD of energy and force predictions with a committee of models
-
-    Uses a compressed data file to keep the extracted information about each trajectory for later
-    """
-
-    verbose = ctx.obj["verbose"]
-
-    inputs = sorted(inputs)
-    if verbose:
-        print(f"input files, count: {len(inputs)}")
-        pprint(inputs)
-
-    if not enforce_rebuild and os.path.isfile(data_fn):
-        # read back
-        data = np.atleast_1d(np.load(data_fn, allow_pickle=True))[0]
-    else:
-        data = trajectory_processing.extract_data_from_trajectories(inputs)
-        np.save(data_fn, data)
-        print("read and saved the data")
-
-    with PdfPages(plot_fn) as pdf:
-        for j, chunk in enumerate(wfl.utils.misc.chunks(inputs, 8)):
-            if verbose:
-                print("plot{:4}".format(j))
-            page_fig = reactions_plotting.plot_one_page_figure(data, chunk)
-            pdf.savefig(page_fig)
-            plt.close(page_fig)
-
-
-@subcli_plot.command("traj-kernel")
-@click.pass_context
-@click.argument("inputs", nargs=-1)
-@click.option("--plot-fn", type=click.STRING, default="plot_kernel.pdf",
-              help="Filename for plot, multi-page")
-def plot_trajectory_std(ctx, inputs, plot_fn):
-    """Plots the kernel similarity of a trajectory with the first frame
-    """
-
-    verbose = ctx.obj["verbose"]
-
-    inputs = sorted(inputs)
-    if verbose:
-        print(f"input files, count: {len(inputs)}")
-        pprint(inputs)
-
-    with PdfPages(plot_fn) as pdf:
-        for j, filename in enumerate(inputs):
-            if verbose:
-                print("plot{:4}".format(j))
-            page_fig = reactions_plotting.calc_and_plot_kernel_vs_first(filename, filename,
-                                                                        index="::8")
-            pdf.savefig(page_fig)
-            plt.close(page_fig)
-
-@subcli_plot.command('normal-modes')
-@click.pass_context
-@click.option('--file-1', '-f1', required=True, type=click.Path(exists=True),
-              help = 'xyz with normal mode information')
-@click.option('--prefix-1', '-p1', required=True, help='prefix for normal modes properties in xyz file')
-@click.option('--label-1', '-l1', help='axis label')
-@click.option('--file-2', '-f2', required=True, type=click.Path(exists=True),
-              help = 'xyz with normal mode information')
-@click.option('--prefix-2', '-p2', required=True, help='prefix for normal modes properties in xyz file')
-@click.option('--label-2', '-l2', help='axis label')
-@click.option('--plot-fn', default='eigenvector_plot.png', show_default=True,
-              help='figure filename')
-@click.option('--arrange-by', type=click.Choice(['frequency', 'order']),
-              help='heatmap with eigenvectors arranged in order or a scatter plot of '
-                   'frequency vs frequency', default='frequency',
-              show_default=True)
-@click.option('--cmap', default='darkred', show_default=True,
-              help='matplotlib color or colormap name')
-@click.option('--adjust-color-range', is_flag=True,
-              help='pick colormap range based on dot product values')
-@click.option('--tolerance', default=1e-4, type=click.FLOAT,
-              show_default=True,
-              help='allowance for dot products to fall outside [0, 1] range')
-def compare_normal_modes(ctx, file_1, prefix_1, file_2, prefix_2, label_1, label_2,
-                         plot_fn, arrange_by, cmap, adjust_color_range, tolerance):
-    """Plot eigenvector dot products and frequencies"""
-
-    normal_modes.eigenvector_plot(nm1=file_1, nm2=file_2, prop_prefix1=prefix_1,
-                                  prop_prefix2=prefix_2, label1=label_1,
-                                  label2=label_2, fig_fname=plot_fn,
-                                  arrange_by=arrange_by, cmap=cmap,
-                                  adjust_color_range=adjust_color_range,
-                                  tolerance=tolerance)
 
 @subcli_generate_configs.command("repeat-buildcell")
 @click.pass_context
